+++ CHANGES.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.3.0] - 2023-06-18
+### Added
+
+- Facial recognition using CompreFace
+
+### Changed
+
+### Fixed
 
 ## [2.2.3] - 2023-04-26
 ### Added